requirements.txt
streamlit==1.29.0
python-dotenv==1.0.0
pyautogen==0.2.0
groq==0.4.0
requests==2.31.0
graphviz2drawio==1.0.0